streaming-project-aws/generator/requirements.txt
kafka-python
pytz==2021.1